FIX: empty medical files field
Fixed a bug where the medical file returned by the /patients/<id>/file
API woold return a medical file with empty allergy and illness arrays
even when they were supposed to hold data.

This bug has now been fixed so that the medical file is now looked up
in the database to get a complete copy.